feat: replace mock_data.py for json files
- Added a new section for reviewing answers after evaluation completion.
- Created a carousel for navigating through reviewed questions.
- Enhanced UI with styles for answer review items, including correct and incorrect answers.
- Updated component logic to handle answer review data and navigation.
- Refactored existing code to maintain consistency in naming conventions (e.g., changed `correctAnswer` to `correct_answer`).
- Modified evaluation result interface to use snake_case for properties.
- Improved user experience with responsive design adjustments for answer review section.

+++ backend/app/models/diagnostic.py
@@ -37,7 +37,7 @@ class EvaluationResult(BaseModel):
     """Final evaluation result"""
     level: DifficultyLevel = Field(..., description="Determined user level")
     score: int = Field(..., ge=0, le=100, description="Overall score percentage")
-    topics: Dict[str, int] = Field(..., description="Score per topic (percentage)")
+    topics: Dict[str, float] = Field(..., description="Score per topic (percentage)")
     learning_style: str = Field(..., description="Determined learning style")
     recommendations: List[str] = Field(..., description="Personalized recommendations")
